+++ packages/css-in-js/package.json
@@ -47,16 +47,16 @@
     "@rollup/plugin-json": "~4.1.0",
     "@rollup/plugin-node-resolve": "~13.1.3",
     "@rollup/plugin-typescript": "~8.3.4",
-    "@types/jest": "~29.5.0",
+    "@types/jest": "~29.5.12",
     "@types/stylis": "^4.0.2",
     "bump": "workspace:~",
     "eslint": "~8.38.0",
-    "jest": "~29.5.0",
+    "jest": "~29.7.0",
     "lint-all": "workspace:~",
     "prettier": "~2.8.7",
     "rollup": "~2.67.3",
     "rollup-plugin-terser": "~7.0.2",
-    "ts-jest": "~29.1.0",
+    "ts-jest": "~29.2.4",
     "typedoc": "~0.25.13",
     "typescript": "~5.5.4"
   },
