+++ packages/fuselage-hooks/package.json
@@ -51,15 +51,15 @@
     "@rollup/plugin-node-resolve": "~13.1.3",
     "@rollup/plugin-typescript": "~8.3.4",
     "@testing-library/react-hooks": "~8.0.1",
-    "@testing-library/user-event": "~14.4.3",
-    "@types/jest": "~29.5.0",
+    "@testing-library/user-event": "~14.5.2",
+    "@types/jest": "~29.5.12",
     "@types/react": "~17.0.80",
     "@types/react-dom": "~17.0.25",
     "@types/resize-observer-browser": "~0.1.7",
     "@types/use-sync-external-store": "~0.0.3",
     "bump": "workspace:~",
     "eslint": "~8.38.0",
-    "jest": "~29.5.0",
+    "jest": "~29.7.0",
     "lint-all": "workspace:~",
     "npm-run-all": "^4.1.5",
     "prettier": "~2.8.7",
@@ -68,7 +68,7 @@
     "rollup": "~2.67.3",
     "rollup-plugin-terser": "~7.0.2",
     "testing-utils": "workspace:~",
-    "ts-jest": "~29.1.0",
+    "ts-jest": "~29.2.4",
     "typedoc": "~0.25.13",
     "typescript": "~5.5.4"
   },
