+++ packages/fuselage-toastbar/package.json
@@ -57,14 +57,15 @@
     "@storybook/react": "~8.2.7",
     "@storybook/react-webpack5": "~8.2.7",
     "@storybook/theming": "~8.2.7",
+    "@testing-library/jest-dom": "~6.4.8",
     "@testing-library/react": "~16.0.0",
     "@testing-library/user-event": "~14.5.2",
-    "@types/jest": "~29.5.0",
+    "@types/jest": "~29.5.12",
     "@types/react": "~17.0.80",
     "@types/react-dom": "~17.0.25",
     "bump": "workspace:~",
     "eslint": "~8.38.0",
-    "jest": "~29.5.0",
+    "jest": "~29.7.0",
     "lint-all": "workspace:~",
     "npm-run-all": "^4.1.5",
     "prettier": "~2.8.7",
@@ -72,7 +73,7 @@
     "rimraf": "~3.0.2",
     "storybook": "~8.2.7",
     "storybook-dark-mode": "~4.0.2",
-    "ts-jest": "~29.1.0",
+    "ts-jest": "~29.2.4",
     "typedoc": "~0.25.13",
     "typescript": "~5.5.4"
   },
@@ -91,21 +92,6 @@
     }
   },
   "prettier": "@rocket.chat/prettier-config/fuselage",
-  "jest": {
-    "preset": "ts-jest",
-    "errorOnDeprecated": true,
-    "testMatch": [
-      "<rootDir>/src/**/*.spec.[jt]s?(x)"
-    ],
-    "testEnvironment": "jsdom",
-    "setupFilesAfterEnv": [
-      "@testing-library/jest-dom/extend-expect",
-      "testing-utils/setup/noErrorsLogged"
-    ],
-    "moduleNameMapper": {
-      "^react-dom/client$": "react-dom"
-    }
-  },
   "volta": {
     "extends": "../../package.json"
   },
