+++ packages/fuselage-toastbar/tsconfig.cjs.json
@@ -6,5 +6,11 @@
     "outDir": "./dist/cjs"
   },
   "include": ["./src"],
-  "exclude": ["**/*.spec.ts", "**/*.stories.tsx"]
+  "exclude": [
+    "**/*.spec.ts",
+    "**/*.spec.tsx",
+    "**/*.stories.tsx",
+    "./jest.config.ts",
+    "./jest-setup.ts"
+  ]
 }
