+++ packages/fuselage-toastbar/tsconfig.esm.json
@@ -6,5 +6,11 @@
     "outDir": "./dist/esm"
   },
   "include": ["./src"],
-  "exclude": ["**/*.spec.ts", "**/*.stories.tsx"]
+  "exclude": [
+    "**/*.spec.ts",
+    "**/*.spec.tsx",
+    "**/*.stories.tsx",
+    "./jest.config.ts",
+    "./jest-setup.ts"
+  ]
 }
