--- packages/fuselage/.jest/global.d.ts
@@ -1,10 +1,0 @@
-export {};
-
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace jest {
-    interface Matchers<R> {
-      toHaveCssInJsClass(): R;
-    }
-  }
-}
