--- packages/fuselage/jest.config.js
@@ -1,19 +1,0 @@
-module.exports = {
-  collectCoverageFrom: [
-    '<rootDir>/src/components/**/*.{js,jsx,ts,tsx}',
-    '!**/*.stories.{js,jsx,ts,tsx}',
-  ],
-  preset: 'ts-jest/presets/js-with-babel',
-  errorOnDeprecated: true,
-  testMatch: ['<rootDir>/src/**/*.spec.{ts,tsx}'],
-  testEnvironment: 'jsdom',
-  moduleNameMapper: {
-    '\\.scss$': 'testing-utils/lazySingletonStyleTagModule',
-    '^react-dom/client$': 'react-dom',
-  },
-  setupFilesAfterEnv: [
-    '@testing-library/jest-dom/extend-expect',
-    'testing-utils/setup/noErrorsLogged',
-    '<rootDir>/.jest/setup.ts',
-  ],
-};
