+++ packages/fuselage/jest.config.ts
@@ -1,0 +1,16 @@
+import type { Config } from 'jest';
+
+export default {
+  preset: 'ts-jest',
+  errorOnDeprecated: true,
+  testMatch: ['<rootDir>/src/**/*.spec.[jt]s?(x)'],
+  testEnvironment: 'jsdom',
+  setupFilesAfterEnv: [
+    '<rootDir>/jest-setup.ts',
+    'testing-utils/setup/noErrorsLogged',
+  ],
+  moduleNameMapper: {
+    '\\.scss$': 'testing-utils/lazySingletonStyleTagModule',
+    '^react-dom/client$': 'react-dom',
+  },
+} satisfies Config;
