+++ packages/fuselage/package.json
@@ -86,12 +86,12 @@
     "@storybook/react-webpack5": "~8.2.7",
     "@storybook/test": "~8.2.7",
     "@storybook/theming": "~8.2.7",
-    "@testing-library/jest-dom": "~5.16.5",
+    "@testing-library/jest-dom": "~6.4.8",
     "@testing-library/react": "~16.0.0",
     "@testing-library/user-event": "~14.5.2",
-    "@types/invariant": "^2.2.35",
-    "@types/jest": "~29.5.0",
-    "@types/jest-axe": "~3.5.5",
+    "@types/invariant": "^2.2.37",
+    "@types/jest": "~29.5.12",
+    "@types/jest-axe": "~3.5.9",
     "autoprefixer": "~10.4.14",
     "babel-loader": "~9.1.2",
     "bump": "workspace:~",
@@ -103,7 +103,7 @@
     "eslint": "~8.38.0",
     "eslint-mdx": "~2.0.5",
     "eslint-plugin-mdx": "~2.0.5",
-    "jest": "~29.5.0",
+    "jest": "~29.7.0",
     "jest-axe": "~8.0.0",
     "lint-all": "workspace:~",
     "mini-css-extract-plugin": "~2.7.6",
@@ -133,7 +133,7 @@
     "stylelint-prettier": "~3.0.0",
     "stylelint-scss": "~4.6.0",
     "testing-utils": "workspace:~",
-    "ts-jest": "~29.1.0",
+    "ts-jest": "~29.2.4",
     "ts-loader": "~9.4.2",
     "typescript": "~5.5.4",
     "webpack": "~5.78.0",
