+++ packages/fuselage/tsconfig.json
@@ -14,6 +14,11 @@
     "allowJs": true,
     "jsx": "react-jsx"
   },
-  "include": ["./src", "./.storybook/**/*", "./.jest/**/*"],
+  "include": [
+    "./src",
+    "./.storybook/**/*",
+    "./jest.config.ts",
+    "./jest-setup.ts"
+  ],
   "exclude": ["./dist", "./storybook-static/", "./*.js"]
 }
