+++ packages/logo/package.json
@@ -45,20 +45,20 @@
     "@rocket.chat/eslint-config-alt": "workspace:~",
     "@rocket.chat/fuselage-tokens": "workspace:~",
     "@rocket.chat/prettier-config": "workspace:~",
-    "@types/jest": "~29.5.0",
+    "@types/jest": "~29.5.12",
     "@types/react": "~17.0.80",
     "@types/react-dom": "~17.0.25",
     "build-logo": "workspace:~",
     "bump": "workspace:~",
     "eslint": "~8.38.0",
-    "jest": "~29.5.0",
+    "jest": "~29.7.0",
     "lint-all": "workspace:~",
     "npm-run-all": "^4.1.5",
     "prettier": "~2.8.7",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "rimraf": "^3.0.2",
-    "ts-jest": "~29.1.0",
+    "ts-jest": "~29.2.4",
     "typedoc": "~0.25.13",
     "typescript": "~5.5.4"
   },
