+++ packages/memo/package.json
@@ -42,14 +42,14 @@
   "devDependencies": {
     "@rocket.chat/eslint-config-alt": "workspace:~",
     "@rocket.chat/prettier-config": "workspace:~",
-    "@types/jest": "~29.5.0",
+    "@types/jest": "~29.5.12",
     "bump": "workspace:~",
     "eslint": "~8.38.0",
-    "jest": "~29.5.0",
+    "jest": "~29.7.0",
     "lint-all": "workspace:~",
     "prettier": "~2.8.7",
     "rimraf": "~3.0.2",
-    "ts-jest": "~29.1.0",
+    "ts-jest": "~29.2.4",
     "typedoc": "~0.25.13",
     "typescript": "~5.5.4"
   },
