+++ packages/mp3-encoder/package.json
@@ -49,15 +49,15 @@
     "@rollup/plugin-commonjs": "~24.1.0",
     "@rollup/plugin-node-resolve": "~15.0.2",
     "@rollup/plugin-typescript": "~11.1.0",
-    "@types/jest": "~29.5.0",
+    "@types/jest": "~29.5.12",
     "bump": "workspace:~",
     "eslint": "~8.38.0",
-    "jest": "~29.5.0",
+    "jest": "~29.7.0",
     "jest-environment-jsdom": "~29.5.0",
     "lint-all": "workspace:~",
     "prettier": "~2.8.7",
     "rollup": "~3.20.4",
-    "ts-jest": "~29.1.0",
+    "ts-jest": "~29.2.4",
     "typedoc": "~0.25.13",
     "typescript": "~5.5.4"
   }
