+++ packages/onboarding-ui/package.json
@@ -58,13 +58,13 @@
     "@storybook/react": "~8.2.7",
     "@storybook/react-webpack5": "~8.2.7",
     "@storybook/theming": "~8.2.7",
-    "@types/jest": "~29.5.0",
+    "@types/jest": "~29.5.12",
     "@types/react": "~17.0.80",
     "@types/react-dom": "~17.0.25",
     "bump": "workspace:~",
     "countries-list": "^2.6.1",
     "eslint": "~8.38.0",
-    "jest": "~29.5.0",
+    "jest": "~29.7.0",
     "lint-all": "workspace:~",
     "npm-run-all": "^4.1.5",
     "prettier": "~2.8.7",
@@ -74,7 +74,7 @@
     "rimraf": "^3.0.2",
     "storybook": "~8.2.7",
     "storybook-dark-mode": "~4.0.2",
-    "ts-jest": "~29.1.0",
+    "ts-jest": "~29.2.4",
     "typedoc": "~0.25.13",
     "typescript": "~5.5.4"
   },
