+++ packages/peggy-loader/package.json
@@ -54,7 +54,7 @@
     "peggy": "3.0.2",
     "prettier": "~2.8.7",
     "rimraf": "^3.0.2",
-    "ts-jest": "~29.1.0",
+    "ts-jest": "~29.2.4",
     "typescript": "~5.5.4",
     "webpack": "~5.78.0"
   }
